+++ packages/bluesky/lib/src/entities/grouped_notification.g.dart
@@ -14,8 +14,11 @@ _$_GroupedNotification _$$_GroupedNotificationFromJson(Map json) =>
       json,
       ($checkedConvert) {
         final val = _$_GroupedNotification(
-          uri: $checkedConvert(
-              'uri', (v) => atUriConverter.fromJson(v as String)),
+          uris: $checkedConvert(
+              'uris',
+              (v) => (v as List<dynamic>)
+                  .map((e) => atUriConverter.fromJson(e as String))
+                  .toList()),
           authors: $checkedConvert(
               'authors',
               (v) => (v as List<dynamic>)
@@ -50,7 +53,7 @@ _$_GroupedNotification _$$_GroupedNotificationFromJson(Map json) =>
 Map<String, dynamic> _$$_GroupedNotificationToJson(
     _$_GroupedNotification instance) {
   final val = <String, dynamic>{
-    'uri': atUriConverter.toJson(instance.uri),
+    'uris': instance.uris.map(atUriConverter.toJson).toList(),
     'authors': instance.authors.map((e) => e.toJson()).toList(),
     'reason': _$GroupedNotificationReasonEnumMap[instance.reason]!,
   };